fix: move hooks folder
