+++ components/molecules/Tooltip.jsx
@@ -1,7 +1,7 @@
 import Image from "next/image";
 import { Tooltip as ReactTooltip } from "react-tooltip";
 import arrow from "@/public/arrow.svg";
-import { usePokemonDescription } from "@/pages/hooks/usePokemonDescription";
+import { usePokemonDescription } from "@/hooks/usePokemonDescription";
 
 export const Tooltip = ({ name, imageUrl }) => {
   const { description } = usePokemonDescription(name);
